+++ examples/pretty-colored.rs
@@ -6,6 +6,8 @@
 //!   line is white
 //! - when the log level is debug, the whole line is white
 //! - when the log level is trace, the whole line is gray ("bright black")
+use std::time::SystemTime;
+
 use fern::colors::{Color, ColoredLevelConfig};
 use log::{debug, error, info, trace, warn};
 
@@ -53,17 +55,17 @@ fn set_up_logging() {
     // configure colors for the name of the level.
     // since almost all of them are the same as the color for the whole line, we
     // just clone `colors_line` and overwrite our changes
-    let colors_level = colors_line.clone().info(Color::Green);
+    let colors_level = colors_line.info(Color::Green);
     // here we set up our fern Dispatch
     fern::Dispatch::new()
         .format(move |out, message, record| {
             out.finish(format_args!(
-                "{color_line}[{date}][{target}][{level}{color_line}] {message}\x1B[0m",
+                "{color_line}[{date} {level} {target} {color_line}] {message}\x1B[0m",
                 color_line = format_args!(
                     "\x1B[{}m",
                     colors_line.get_color(&record.level()).to_fg_str()
                 ),
-                date = chrono::Local::now().format("%Y-%m-%d %H:%M:%S"),
+                date = humantime::format_rfc3339_seconds(SystemTime::now()),
                 target = record.target(),
                 level = colors_level.color(record.level()),
                 message = message,