+++ src/lib.rs
@@ -1,6 +1,17 @@
 #![deny(missing_docs)]
-#![doc(html_root_url = "https://docs.rs/fern/0.6.1")]
+#![doc(html_root_url = "https://docs.rs/fern/0.7.1")]
 //! Efficient, configurable logging in Rust.
+//!
+//! # fern 0.4.4, 0.5.\*, 0.6.\* security warning - `colored` feature + global allocator
+//!
+//! One of our downstream dependencies, [atty](https://docs.rs/atty/), through
+//! [colored](https://docs.rs/colored/), has an unsoundness issue:
+//! <https://rustsec.org/advisories/RUSTSEC-2021-0145.html>.
+//!
+//! This shows up in one situation: if you're using `colored` 0.1.0 (the crate, or our
+//! feature), and a custom global allocator.
+//!
+//! Upgrade to `fern` 0.7.0, and `colored` 0.2.0 if you depend on it directly, to fix this issue.
 //!
 //! # Depending on fern
 //!
@@ -9,7 +20,7 @@
 //! ```toml
 //! [dependencies]
 //! log = "0.4"
-//! fern = "0.6"
+//! fern = "0.7"
 //! ```
 //!
 //! # Example setup
@@ -22,15 +33,16 @@
 //!
 //! ```no_run
 //! use log::{debug, error, info, trace, warn};
+//! use std::time::SystemTime;
 //!
 //! fn setup_logger() -> Result<(), fern::InitError> {
 //!     fern::Dispatch::new()
 //!         .format(|out, message, record| {
 //!             out.finish(format_args!(
-//!                 "{}[{}][{}] {}",
-//!                 chrono::Local::now().format("[%Y-%m-%d][%H:%M:%S]"),
-//!                 record.target(),
+//!                 "[{} {} {}] {}",
+//!                 humantime::format_rfc3339_seconds(SystemTime::now()),
 //!                 record.level(),
+//!                 record.target(),
 //!                 message
 //!             ))
 //!         })
@@ -40,96 +52,134 @@
 //!         .apply()?;
 //!     Ok(())
 //! }
-//! # fn main() {
-//! #     setup_logger().expect("failed to set up logger")
-//! # }
+//!
+//! fn main() -> Result<(), Box<dyn std::error::Error>> {
+//!     setup_logger()?;
+//!
+//!     info!("Hello, world!");
+//!     warn!("Warning!");
+//!     debug!("Now exiting.");
+//!
+//!     Ok(())
+//! }
 //! ```
 //!
 //! Let's unwrap this:
 //!
-//! ---
 //!
-//! [`fern::Dispatch::new()`]
+//! ```
+//! fern::Dispatch::new()
+//! # ;
+//! ```
 //!
-//! Create an empty configuration.
+//! [`Dispatch::new`] creates an empty configuration.
 //!
-//! ---
+//! ```
+//! # fern::Dispatch::new()
+//! .format(|out, message, record| {
+//!     out.finish(format_args!(
+//!         "..."
+//!     ))
+//! })
+//! # ;
+//! ```
 //!
-//! [`.format(|...| ...)`]
+//! This incantation sets the `Dispatch` format! The closure taking in
+//! `out, message, record` will be called once for each message going through
+//! the dispatch, and the formatted log message will be used for any downstream
+//! consumers.
 //!
-//! Add a formatter to the logger, modifying all messages sent through.
+//! Do any work you want in this closure, and then call `out.finish` at the end.
+//! The callback-style result passing with `out.finish(format_args!())` lets us
+//! format without any intermediate string allocation.
 //!
-//! ___
+//! [`format_args!`] has the same format as [`println!`], just returning a
+//! not-yet-written result we can use internally.
 //!
-//! [`chrono::Local::now()`]
+//! ```
+//! std::time::SystemTime::now()
+//! # ;
+//! ```
 //!
-//! Get the current time in the local timezone using the [`chrono`] library.
-//! See the [time-and-date docs].
+//! [`std::time::SystemTime::now`] retrieves the current time.
 //!
-//! ___
+//! ```
+//! humantime::format_rfc3339_seconds(
+//!     // ...
+//!     # std::time::SystemTime::now()
+//! )
+//! # ;
+//! ```
 //!
-//! [`.format("[%Y-%m-%d][%H:%M:%S]")`][chrono-format]
+//! [`humantime::format_rfc3339_seconds`] formats the current time into an
+//! RFC3339 timestamp, with second-precision.
 //!
-//! Use chrono's lazy format specifier to turn the time into a readable string.
+//! RFC3339 looks like `2018-02-14T00:28:07Z`, always using UTC, ignoring system
+//! timezone.
 //!
-//! ---
+//! `humantime` is a nice light dependency, but only offers this one format.
+//! For more custom time formatting, I recommend
+//! [`jiff`](https://docs.rs/jiff/).
 //!
-//! [`out.finish(format_args!(...))`]
+//! Now, back to the [`Dispatch`] methods:
 //!
-//! Call the `fern::FormattingCallback` to submit the formatted message.
+//! ```
+//! # fern::Dispatch::new()
+//! .level(log::LevelFilter::Debug)
+//! # ;
+//! ```
 //!
-//! This roundabout way is slightly odd, but it allows for very fast logging.
-//! No string allocation required!
+//! Sets the minimum logging level for all modules, if not overwritten with
+//! [`Dispatch::level_for`], to [`Level::Debug`][log::Level::Debug].
 //!
-//! [`format_args!()`] has the same format as [`println!()`] \(and every other
-//! [`std::fmt`]-based macro).
+//! ```
+//! # fern::Dispatch::new()
+//! .chain(std::io::stdout())
+//! # ;
+//! ```
 //!
-//! ---
-//!
-//! [`.level(log::LevelFilter::Debug)`]
-//!
-//! Set the minimum level needed to output to `Debug`.
-//!
-//! ---
-//!
-//! [`.chain(std::io::stdout())`]
-//!
-//! Add a child to the logger. All messages which pass the filters will be sent
-//! to stdout.
+//! Adds a child to the logger. With this, all messages which pass the filters
+//! will be sent to stdout.
 //!
 //! [`Dispatch::chain`] accepts [`Stdout`], [`Stderr`], [`File`] and other
 //! [`Dispatch`] instances.
 //!
-//! ---
+//! ```no_run
+//! # fern::Dispatch::new()
+//! .chain(fern::log_file("output.log")?)
+//! # ; <Result<(), Box<dyn std::error::Error>>>::Ok(())
+//! ```
 //!
-//! [`.chain(fern::log_file(...)?)`]
+//! Adds a second child sending messages to the file "output.log".
 //!
-//! Add a second child sending messages to the file "output.log".
+//! See [`log_file`].
 //!
-//! See [`fern::log_file()`] for more info on file output.
+//! ```
+//! # fern::Dispatch::new()
+//! // ...
+//! .apply()
+//! # ;
+//! ```
 //!
-//! ---
-//!
-//! [`.apply()`][`.apply`]
-//!
-//! Consume the configuration and instantiate it as the current runtime global
+//! Consumes the configuration and instantiates it as the current runtime global
 //! logger.
 //!
 //! This will fail if and only if `.apply()` or equivalent form another crate
 //! has already been used this runtime.
 //!
-//! Since the binary crate is the only one ever setting up logging, the
-//! [`apply`] result can be reasonably unwrapped: it's a bug if any crate is
-//! calling this method more than once.
+//! Since the binary crate is the only one ever setting up logging, and it's
+//! usually done near the start of `main`, the [`Dispatch::apply`] result can be
+//! reasonably unwrapped: it's a bug if any crate is calling this method more
+//! than once.
 //!
 //! ---
 //!
 //! The final output will look like:
 //!
 //! ```text
-//! [2017-01-20][12:55:04][crate-name][INFO] Hello, world!
-//! [2017-01-20][12:56:21][crate-name][WARN] Ahhh!
-//! [2017-01-20][12:58:00][crate-name][DEBUG] Something less important happened.
+//! [2023-03-18T20:12:50Z INFO cmd_program] Hello, world!
+//! [2023-03-18T20:12:50Z WARN cmd_program] Warning!
+//! [2023-03-18T20:12:50Z DEBUG cmd_program] Now exiting.
 //! ```
 //!
 //! # Logging
@@ -159,7 +209,7 @@
 //!
 //! # More
 //!
-//! The [`Dispatch` documentation] has example usages of each method, and the
+//! The [`Dispatch`] documentation has example usages of each method, and the
 //! [full example program] might be useful for using fern in a larger
 //! application context.
 //!
@@ -171,34 +221,12 @@
 //! See the [meta] module for information on getting logging-within-logging
 //! working correctly.
 //!
-//! [`fern::Dispatch::new()`]: struct.Dispatch.html#method.new
-//! [`.format(|...| ...)`]: struct.Dispatch.html#method.format
-//! [`chrono::Local::now()`]: https://docs.rs/chrono/0.4/chrono/offset/local/struct.Local.html#method.now
-//! [chrono-format]: https://docs.rs/chrono/0.4/chrono/datetime/struct.DateTime.html#method.format
-//! [`out.finish(format_args!(...))`]: struct.FormatCallback.html#method.finish
-//! [`.level(log::LevelFilter::Debug)`]: struct.Dispatch.html#method.level
-//! [`Dispatch::chain`]: struct.Dispatch.html#method.chain
-//! [`.chain(std::io::stdout())`]: struct.Dispatch.html#method.chain
-//! [`Stdout`]: https://doc.rust-lang.org/std/io/struct.Stdout.html
-//! [`Stderr`]: https://doc.rust-lang.org/std/io/struct.Stderr.html
-//! [`File`]: https://doc.rust-lang.org/std/fs/struct.File.html
-//! [`Dispatch`]: struct.Dispatch.html
-//! [`.chain(fern::log_file(...)?)`]: struct.Dispatch.html#method.chain
-//! [`fern::log_file()`]: fn.log_file.html
-//! [`.apply`]: struct.Dispatch.html#method.apply
-//! [`format_args!()`]: https://doc.rust-lang.org/std/macro.format_args.html
-//! [`println!()`]: https://doc.rust-lang.org/std/macro.println.html
-//! [`std::fmt`]: https://doc.rust-lang.org/std/fmt/
-//! [`chrono`]: https://github.com/chronotope/chrono
-//! [time-and-date docs]: https://docs.rs/chrono/0.4/chrono/index.html#date-and-time
-//! [the format specifier docs]: https://docs.rs/chrono/0.4/chrono/format/strftime/index.html#specifiers
-//! [`Dispatch` documentation]: struct.Dispatch.html
-//! [full example program]: https://github.com/daboross/fern/tree/fern-0.6.1/examples/cmd-program.rs
-//! [syslog full example program]: https://github.com/daboross/fern/tree/fern-0.6.1/examples/syslog.rs
-//! [`apply`]: struct.Dispatch.html#method.apply
-//! [colors]: colors/index.html
-//! [syslog]: syslog/index.html
-//! [meta]: meta/index.html
+//! [`Stdout`]: std::io::Stdout
+//! [`Stderr`]: std::io::Stderr
+//! [`File`]: std::fs::File
+//! [full example program]: https://github.com/daboross/fern/tree/fern-0.7.0/examples/cmd-program.rs
+//! [syslog full example program]: https://github.com/daboross/fern/tree/fern-0.7.0/examples/syslog.rs
+//! [`humantime::format_rfc3339_seconds`]: https://docs.rs/humantime/2/humantime/fn.format_rfc3339_seconds.html
 use std::{
     convert::AsRef,
     fmt,
@@ -207,8 +235,11 @@ use std::{
     path::Path,
 };
 
-#[cfg(all(not(windows), feature = "syslog-4"))]
+#[cfg(all(not(windows), any(feature = "syslog-4", feature = "syslog-6")))]
 use std::collections::HashMap;
+
+#[cfg(all(not(windows), feature = "syslog-7"))]
+use std::collections::BTreeMap;
 
 pub use crate::{
     builders::{Dispatch, Output, Panic},
@@ -225,7 +256,14 @@ mod log_impl;
 
 #[cfg(feature = "colored")]
 pub mod colors;
-#[cfg(all(not(windows), feature = "syslog-3", feature = "syslog-4"))]
+#[cfg(all(
+    feature = "syslog-3",
+    feature = "syslog-4",
+    // disable on windows when running doctests, as the code itself only runs on
+    // linux. enable on windows otherwise because it's a documentation-only
+    // module.
+    any(not(windows), not(doctest))
+))]
 pub mod syslog;
 
 pub mod meta;
@@ -262,6 +300,24 @@ type Syslog6Rfc3164Logger = syslog6::Logger<syslog6::LoggerBackend, syslog6::For
 #[cfg(all(not(windows), feature = "syslog-6"))]
 type Syslog6Rfc5424Logger = syslog6::Logger<syslog6::LoggerBackend, syslog6::Formatter5424>;
 
+#[cfg(all(not(windows), feature = "syslog-7"))]
+type Syslog7Rfc3164Logger = syslog7::Logger<syslog7::LoggerBackend, syslog7::Formatter3164>;
+
+#[cfg(all(not(windows), feature = "syslog-7"))]
+type Syslog7Rfc5424Logger = syslog7::Logger<syslog7::LoggerBackend, syslog7::Formatter5424>;
+
+#[cfg(all(not(windows), feature = "syslog-4"))]
+type Syslog4TransformFn =
+    dyn Fn(&log::Record) -> (i32, HashMap<String, HashMap<String, String>>, String) + Send + Sync;
+
+#[cfg(all(not(windows), feature = "syslog-6"))]
+type Syslog6TransformFn =
+    dyn Fn(&log::Record) -> (u32, HashMap<String, HashMap<String, String>>, String) + Send + Sync;
+
+#[cfg(all(not(windows), feature = "syslog-7"))]
+type Syslog7TransformFn =
+    dyn Fn(&log::Record) -> (u32, BTreeMap<String, BTreeMap<String, String>>, String) + Send + Sync;
+
 /// Convenience method for opening a log file with common options.
 ///
 /// Equivalent to:
@@ -280,11 +336,7 @@ type Syslog6Rfc5424Logger = syslog6::Logger<syslog6::LoggerBackend, syslog6::For
 /// [`OpenOptions`]: https://doc.rust-lang.org/std/fs/struct.OpenOptions.html
 #[inline]
 pub fn log_file<P: AsRef<Path>>(path: P) -> io::Result<File> {
-    OpenOptions::new()
-        .write(true)
-        .create(true)
-        .append(true)
-        .open(path)
+    OpenOptions::new().create(true).append(true).open(path)
 }
 
 /// Convenience method for opening a re-openable log file with common options.
@@ -313,9 +365,7 @@ pub fn log_reopen(path: &Path, signal: Option<libc::c_int>) -> io::Result<reopen
     let r = reopen03::Reopen::new(Box::new(move || log_file(&p)))?;
 
     if let Some(s) = signal {
-        if let Err(e) = r.handle().register_signal(s) {
-            return Err(e);
-        }
+        r.handle().register_signal(s)?;
     }
     Ok(r)
 }
@@ -340,9 +390,10 @@ pub fn log_reopen(path: &Path, signal: Option<libc::c_int>) -> io::Result<reopen
 /// This function requires the `reopen-1` feature to be enabled.
 #[cfg(all(not(windows), feature = "reopen-1"))]
 #[inline]
-pub fn log_reopen1<S: IntoIterator<Item = libc::c_int>>(path: &Path, signals: S)
-    -> io::Result<reopen1::Reopen<File>>
-{
+pub fn log_reopen1<S: IntoIterator<Item = libc::c_int>>(
+    path: &Path,
+    signals: S,
+) -> io::Result<reopen1::Reopen<File>> {
     let p = path.to_owned();
     let r = reopen1::Reopen::new(Box::new(move || log_file(&p)))?;