+++ src/log_impl.rs
@@ -18,13 +18,11 @@ use log::{self, Log};
 use crate::{Filter, Formatter};
 
 #[cfg(all(not(windows), feature = "syslog-4"))]
-use crate::{Syslog4Rfc3164Logger, Syslog4Rfc5424Logger};
+use crate::{Syslog4Rfc3164Logger, Syslog4Rfc5424Logger, Syslog4TransformFn};
 #[cfg(all(not(windows), feature = "syslog-6"))]
-use crate::{Syslog6Rfc3164Logger, Syslog6Rfc5424Logger};
-#[cfg(all(not(windows), feature = "reopen-03"))]
-use reopen03;
-#[cfg(all(not(windows), feature = "reopen-1"))]
-use reopen1;
+use crate::{Syslog6Rfc3164Logger, Syslog6Rfc5424Logger, Syslog6TransformFn};
+#[cfg(all(not(windows), feature = "syslog-7"))]
+use crate::{Syslog7Rfc3164Logger, Syslog7Rfc5424Logger, Syslog7TransformFn};
 
 pub enum LevelConfiguration {
     JustDefault,
@@ -77,6 +75,10 @@ pub enum Output {
     Syslog6Rfc3164(Syslog6Rfc3164),
     #[cfg(all(not(windows), feature = "syslog-6"))]
     Syslog6Rfc5424(Syslog6Rfc5424),
+    #[cfg(all(not(windows), feature = "syslog-7"))]
+    Syslog7Rfc3164(Syslog7Rfc3164),
+    #[cfg(all(not(windows), feature = "syslog-7"))]
+    Syslog7Rfc5424(Syslog7Rfc5424),
     Dispatch(Dispatch),
     SharedDispatch(Arc<Dispatch>),
     OtherBoxed(Box<dyn Log>),
@@ -143,11 +145,7 @@ pub struct Syslog4Rfc3164 {
 #[cfg(all(not(windows), feature = "syslog-4"))]
 pub struct Syslog4Rfc5424 {
     pub inner: Mutex<Syslog4Rfc5424Logger>,
-    pub transform: Box<
-        dyn Fn(&log::Record) -> (i32, HashMap<String, HashMap<String, String>>, String)
-            + Sync
-            + Send,
-    >,
+    pub transform: Box<Syslog4TransformFn>,
 }
 
 #[cfg(all(not(windows), feature = "syslog-6"))]
@@ -158,11 +156,18 @@ pub struct Syslog6Rfc3164 {
 #[cfg(all(not(windows), feature = "syslog-6"))]
 pub struct Syslog6Rfc5424 {
     pub inner: Mutex<Syslog6Rfc5424Logger>,
-    pub transform: Box<
-        dyn Fn(&log::Record) -> (u32, HashMap<String, HashMap<String, String>>, String)
-            + Sync
-            + Send,
-    >,
+    pub transform: Box<Syslog6TransformFn>,
+}
+
+#[cfg(all(not(windows), feature = "syslog-7"))]
+pub struct Syslog7Rfc3164 {
+    pub inner: Mutex<Syslog7Rfc3164Logger>,
+}
+
+#[cfg(all(not(windows), feature = "syslog-7"))]
+pub struct Syslog7Rfc5424 {
+    pub inner: Mutex<Syslog7Rfc5424Logger>,
+    pub transform: Box<Syslog7TransformFn>,
 }
 
 pub struct Panic;
@@ -294,15 +299,11 @@ impl DateBasedConfig {
     }
 
     pub fn open_log_file(path: &Path) -> io::Result<fs::File> {
-        OpenOptions::new()
-            .write(true)
-            .create(true)
-            .append(true)
-            .open(path)
+        OpenOptions::new().create(true).append(true).open(path)
     }
 
     pub fn open_current_log_file(&self, suffix: &str) -> io::Result<fs::File> {
-        Self::open_log_file(&self.compute_file_path(&suffix))
+        Self::open_log_file(&self.compute_file_path(suffix))
     }
 }
 
@@ -407,8 +408,8 @@ impl LevelConfiguration {
             LevelConfiguration::JustDefault => None,
             LevelConfiguration::Minimal(ref levels) => levels
                 .iter()
-                .find(|&&(ref test_module, _)| test_module == module)
-                .map(|&(_, level)| level),
+                .find(|(test_module, _)| test_module == module)
+                .map(|(_, level)| *level),
             LevelConfiguration::Many(ref levels) => levels.get(module).cloned(),
         }
     }
@@ -435,6 +436,10 @@ impl Log for Output {
             Output::Syslog6Rfc3164(ref s) => s.enabled(metadata),
             #[cfg(all(not(windows), feature = "syslog-6"))]
             Output::Syslog6Rfc5424(ref s) => s.enabled(metadata),
+            #[cfg(all(not(windows), feature = "syslog-7"))]
+            Output::Syslog7Rfc3164(ref s) => s.enabled(metadata),
+            #[cfg(all(not(windows), feature = "syslog-7"))]
+            Output::Syslog7Rfc5424(ref s) => s.enabled(metadata),
             Output::Panic(ref s) => s.enabled(metadata),
             Output::Writer(ref s) => s.enabled(metadata),
             #[cfg(feature = "date-based")]
@@ -468,6 +473,10 @@ impl Log for Output {
             Output::Syslog6Rfc3164(ref s) => s.log(record),
             #[cfg(all(not(windows), feature = "syslog-6"))]
             Output::Syslog6Rfc5424(ref s) => s.log(record),
+            #[cfg(all(not(windows), feature = "syslog-7"))]
+            Output::Syslog7Rfc3164(ref s) => s.log(record),
+            #[cfg(all(not(windows), feature = "syslog-7"))]
+            Output::Syslog7Rfc5424(ref s) => s.log(record),
             Output::Panic(ref s) => s.log(record),
             Output::Writer(ref s) => s.log(record),
             #[cfg(feature = "date-based")]
@@ -501,6 +510,10 @@ impl Log for Output {
             Output::Syslog6Rfc3164(ref s) => s.flush(),
             #[cfg(all(not(windows), feature = "syslog-6"))]
             Output::Syslog6Rfc5424(ref s) => s.flush(),
+            #[cfg(all(not(windows), feature = "syslog-7"))]
+            Output::Syslog7Rfc3164(ref s) => s.flush(),
+            #[cfg(all(not(windows), feature = "syslog-7"))]
+            Output::Syslog7Rfc5424(ref s) => s.flush(),
             Output::Panic(ref s) => s.flush(),
             Output::Writer(ref s) => s.flush(),
             #[cfg(feature = "date-based")]
@@ -636,7 +649,7 @@ impl Manual {
     }
 }
 
-impl<'a> FormatCallback<'a> {
+impl FormatCallback<'_> {
     /// Complete the formatting call that this FormatCallback was created for.
     ///
     /// This will call the rest of the logging chain using the given formatted
@@ -784,7 +797,15 @@ impl Log for Sender {
     fn flush(&self) {}
 }
 
-#[cfg(all(not(windows), any(feature = "syslog-3", feature = "syslog-4", feature = "syslog-6")))]
+#[cfg(all(
+    not(windows),
+    any(
+        feature = "syslog-3",
+        feature = "syslog-4",
+        feature = "syslog-6",
+        feature = "syslog-7"
+    )
+))]
 macro_rules! send_syslog {
     ($logger:expr, $level:expr, $message:expr) => {
         use log::Level;
@@ -870,6 +891,42 @@ impl Log for Syslog6Rfc3164 {
 
 #[cfg(all(not(windows), feature = "syslog-6"))]
 impl Log for Syslog6Rfc5424 {
+    fn enabled(&self, _: &log::Metadata) -> bool {
+        true
+    }
+
+    fn log(&self, record: &log::Record) {
+        fallback_on_error(record, |record| {
+            let transformed = (self.transform)(record);
+            let mut log = self.inner.lock().unwrap_or_else(|e| e.into_inner());
+            send_syslog!(log, record.level(), transformed);
+
+            Ok(())
+        });
+    }
+    fn flush(&self) {}
+}
+
+#[cfg(all(not(windows), feature = "syslog-7"))]
+impl Log for Syslog7Rfc3164 {
+    fn enabled(&self, _: &log::Metadata) -> bool {
+        true
+    }
+
+    fn log(&self, record: &log::Record) {
+        fallback_on_error(record, |record| {
+            let message = record.args().to_string();
+            let mut log = self.inner.lock().unwrap_or_else(|e| e.into_inner());
+            send_syslog!(log, record.level(), message);
+
+            Ok(())
+        });
+    }
+    fn flush(&self) {}
+}
+
+#[cfg(all(not(windows), feature = "syslog-7"))]
+impl Log for Syslog7Rfc5424 {
     fn enabled(&self, _: &log::Metadata) -> bool {
         true
     }
@@ -1038,6 +1095,8 @@ enum LogError {
     Syslog4(syslog4::Error),
     #[cfg(all(not(windows), feature = "syslog-6"))]
     Syslog6(syslog6::Error),
+    #[cfg(all(not(windows), feature = "syslog-7"))]
+    Syslog7(syslog7::Error),
 }
 
 impl fmt::Display for LogError {
@@ -1049,6 +1108,8 @@ impl fmt::Display for LogError {
             LogError::Syslog4(ref e) => write!(f, "{}", e),
             #[cfg(all(not(windows), feature = "syslog-6"))]
             LogError::Syslog6(ref e) => write!(f, "{}", e),
+            #[cfg(all(not(windows), feature = "syslog-7"))]
+            LogError::Syslog7(ref e) => write!(f, "{}", e),
         }
     }
 }
@@ -1076,6 +1137,13 @@ impl From<syslog4::Error> for LogError {
 impl From<syslog6::Error> for LogError {
     fn from(error: syslog6::Error) -> Self {
         LogError::Syslog6(error)
+    }
+}
+
+#[cfg(all(not(windows), feature = "syslog-7"))]
+impl From<syslog7::Error> for LogError {
+    fn from(error: syslog7::Error) -> Self {
+        LogError::Syslog7(error)
     }
 }